cookiecutters/poetry-dash-app-data/{{cookiecutter.app_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.app_name}}"
version = "{{cookiecutter.version}}"
description = "{{cookiecutter.description}}"
authors = ["{{cookiecutter.authors}}"]

[tool.poetry.dependencies]
python = "^3.8"
dash = "^1.9.1"
dash-bootstrap-components = "^0.9.1"
python-box = "^4.2.2"
pydash = "^4.7.6"
hamplify = "^1.3.0"
lesscpy = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyinstaller = "^3.6"
pywin32-ctypes = "^0.2.0"
pefile = "^2019.4.18"

[rhyme.scripts]
run = "python -m {{cookiecutter.app_name}}"
haml = "hamplify {{cookiecutter.app_name}}/dash_app/assets/markup/haml {{cookiecutter.app_name}}/dash_app/assets/markup/html"
less = "lesscpy -r {{cookiecutter.app_name}}/dash_app/assets/styles/less -o {{cookiecutter.app_name}}/dash_app/assets/styles/css"
build = "pyinstaller {{cookiecutter.app_name}}/__main__.py --distpath dist -n {{cookiecutter.app_name}}.exe -F --add-data .venv/Lib/site-packages/dash_html_components;dash_html_components --add-data .venv/Lib/site-packages/dash_core_components;dash_core_components --add-data .venv/Lib/site-packages/dash_renderer;dash_renderer --add-data .venv/Lib/site-packages/dash;dash --add-data .venv/Lib/site-packages/dash_table;dash_table --add-data .venv/Lib/site-packages/dash_bootstrap_components;dash_bootstrap_components --add-data {{cookiecutter.app_name}}/dash_app/assets;assets"
exec = "dist/{{cookiecutter.app_name}}.exe"
build_dir = "pyinstaller {{cookiecutter.app_name}}/__main__.py --distpath dist -n {{cookiecutter.app_name}} --add-data .venv/Lib/site-packages/dash_html_components;dash_html_components --add-data .venv/Lib/site-packages/dash_core_components;dash_core_components --add-data .venv/Lib/site-packages/dash_renderer;dash_renderer --add-data .venv/Lib/site-packages/dash;dash --add-data .venv/Lib/site-packages/dash_table;dash_table --add-data .venv/Lib/site-packages/dash_bootstrap_components;dash_bootstrap_components --add-data {{cookiecutter.app_name}}/dash_app/assets;assets"
exec_dir = "dist/{{cookiecutter.app_name}}/{{cookiecutter.app_name}}.exe"
clear = "rmdir build /s /q & rmdir dist /s /q & del {{cookiecutter.app_name}}.exe.spec & del {{cookiecutter.app_name}}.spec"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"